Add option to check max request size
There is a maximum of 500kb per batch request and 32kb per call. The
API will return a 400 response for batch sizes that are over 500kb but
will still return 200 if the individual calls are over 32kb.
Unfortunately, those calls will silently fail. This option allows PHP
to raise an error when request sizes are over the API set limit to
avoid silently lost data.